Simplify GCE resource lister
