--- pkg/resources/gce/subnet.go
@@ -1,35 +1,0 @@
-/*
-Copyright 2017 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package gce
-
-import (
-	compute "google.golang.org/api/compute/v1"
-	"k8s.io/kops/pkg/resources"
-	"k8s.io/kops/upup/pkg/fi/cloudup/gce"
-)
-
-// DumpSubnetwork is responsible for dumping a resource for a Subnetwork
-func DumpSubnetwork(op *resources.DumpOperation, r *resources.Resource) error {
-	obj := r.Obj.(*compute.Subnetwork)
-
-	subnet := &resources.Subnet{
-		ID: gce.LastComponent(obj.SelfLink),
-	}
-	op.Dump.Subnets = append(op.Dump.Subnets, subnet)
-
-	return nil
-}
